Fix encoding type conflicts after merge: align int types, add yaml.v3 dep
- Remove duplicate encoding.go/encoding_test.go (keep conversation.go)
- Add Content field to Message (db reader uses Content, encoding uses Text)
- Add SenderID to Reaction struct
- Change int64 to int for ID fields (Conversation, Message, Attachment)
- Fix function calls to EncodeConversation (pass pointer, not value)
- Add gopkg.in/yaml.v3 dependency for contextengine
- All tests pass

+++ internal/contextengine/convos.go
@@ -545,7 +545,7 @@ func loadConversationWithMessages(db *sql.DB, convID, chatID int64) (*encoding.C
 }
 
 // loadAttachmentsForMessage loads attachments for a message
-func loadAttachmentsForMessage(db *sql.DB, messageID int64) ([]encoding.Attachment, error) {
+func loadAttachmentsForMessage(db *sql.DB, messageID int) ([]encoding.Attachment, error) {
 	query := `
 		SELECT id, mime_type, file_name, is_sticker
 		FROM attachments
